Format an estimate, lower and upper confidence intervals into a single character string

Take numeric/character vectors of the estimate, and upper/lower confidence/credible intervals, round with trailing zeros and format into the required journal style.

pretty_ci(est, lci, uci, sep = " to ", digits = 2, inline = FALSE,
  note = "95% CI")

Arguments

est
A numeric or character vector of estimates.

lci
A numeric or character vector of lower confidence/credible intervals.

uci
A numeric or character vector of upper confidence/credible intervals.

sep
A character vector indicating the seperator used between the upper and lower confidence/credible intervals. The default is ' to '.

digits
Integer indicating the number of decimal places to be used.

inline
Logical operator indicating whether an explanatory note is required.

note
A character vector indicating the explanatory note to be used.

Value

The estimate with formated upper, and lower confidence/credible intervals as a character vector.

See also

pretty_round pretty_percentage

Examples

## Formating a single confidence interval
pretty_ci(2, lci = 1, uci = 3)

#> [1] "2.00 (1.00 to 3.00)"

## Formating vectors of confidence intervals
est <- c(0, 1, 100, 300, 21221, 403)
lci <- c(-123, -0.2, 50, 100, 12321, 200)
uci <- c(10, 2, 200, 400, 30122, 500)

pretty_ci(est, lci = lci, uci = uci, sep = '-', digits = 1)

#> [1] "0.0 (-123.0-10.0)"         "1.0 (-0.2-2.0)"           
#> [3] "100.0 (50.0-200.0)"        "300.0 (100.0-400.0)"      
#> [5] "21221.0 (12321.0-30122.0)" "403.0 (200.0-500.0)"      

## Use in a dplyr workflow
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

x <- data_frame(est = c(0,1), lci = c(0, 2), uci = c(1, 4))
x <- x %>% mutate(ci = est %>% pretty_ci(lci = lci, uci = uci, sep = ' by ', digits = 0))